refactor: move logic to SessionManager
