+++ android/src/main/kotlin/com/rudderstack/sdk/kotlin/android/plugins/sessiontracking/SessionTrackingPlugin.kt
@@ -1,85 +1,36 @@
 package com.rudderstack.sdk.kotlin.android.plugins.sessiontracking
 
-import com.rudderstack.sdk.kotlin.android.DEFAULT_SESSION_TIMEOUT_IN_MILLIS
-import com.rudderstack.sdk.kotlin.android.plugins.lifecyclemanagment.ActivityLifecycleObserver
-import com.rudderstack.sdk.kotlin.android.plugins.lifecyclemanagment.ProcessLifecycleObserver
-import com.rudderstack.sdk.kotlin.android.utils.addLifecycleObserver
-import com.rudderstack.sdk.kotlin.android.utils.getMonotonicCurrentTime
 import com.rudderstack.sdk.kotlin.android.utils.mergeWithHigherPriorityTo
 import com.rudderstack.sdk.kotlin.core.Analytics
-import com.rudderstack.sdk.kotlin.core.internals.logger.LoggerAnalytics
 import com.rudderstack.sdk.kotlin.core.internals.models.Event
 import com.rudderstack.sdk.kotlin.core.internals.plugins.Plugin
-import com.rudderstack.sdk.kotlin.core.internals.statemanagement.State
-import com.rudderstack.sdk.kotlin.core.internals.storage.Storage
-import com.rudderstack.sdk.kotlin.core.internals.utils.DateTimeUtils
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.launch
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.put
-import java.util.concurrent.TimeUnit
-import kotlin.properties.Delegates
 import com.rudderstack.sdk.kotlin.android.Analytics as AndroidAnalytics
-import com.rudderstack.sdk.kotlin.android.Configuration as AndroidConfiguration
 
 internal const val SESSION_ID = "sessionId"
 internal const val SESSION_START = "sessionStart"
 
 @Suppress("TooManyFunctions")
-@OptIn(ExperimentalCoroutinesApi::class)
-internal class SessionTrackingPlugin(
-    // single thread dispatcher is required so that the session variables are updated (on storage) in a sequential manner.
-    private val sessionDispatcher: CoroutineDispatcher = Dispatchers.IO.limitedParallelism(1)
-) : Plugin {
+internal class SessionTrackingPlugin : Plugin {
 
     override val pluginType: Plugin.PluginType = Plugin.PluginType.PreProcess
     override lateinit var analytics: Analytics
-    private val storage: Storage
-        get() = analytics.storage
-
-    private lateinit var sessionState: State<SessionState>
-
-    internal val sessionId
-        get() = sessionState.value.sessionId
-    private val lastActivityTime
-        get() = sessionState.value.lastActivityTime
-    private val isSessionManual
-        get() = sessionState.value.isSessionManual
-    private val isSessionStart
-        get() = sessionState.value.isSessionStart
-
-    internal var sessionTimeout by Delegates.notNull<Long>()
+    private lateinit var sessionManager: SessionManager
 
     override fun setup(analytics: Analytics) {
         super.setup(analytics)
 
-        (analytics.configuration as? AndroidConfiguration)?.let { config ->
-            sessionTimeout = if (config.sessionConfiguration.sessionTimeoutInMillis >= 0) {
-                config.sessionConfiguration.sessionTimeoutInMillis
-            } else {
-                LoggerAnalytics.error("Session timeout cannot be negative. Setting it to default value.")
-                DEFAULT_SESSION_TIMEOUT_IN_MILLIS
-            }
-            sessionState = State(SessionState.initialState(analytics.storage))
-
-            when {
-                config.sessionConfiguration.automaticSessionTracking -> {
-                    checkAndStartSessionOnLaunch()
-                    attachSessionTrackingObservers()
-                }
-                !isSessionManual -> endSession()
-                else -> Unit
-            }
+        (analytics as? AndroidAnalytics)?.let {
+            sessionManager = analytics.sessionManager
         }
     }
 
     override suspend fun intercept(event: Event): Event {
-        if (sessionId != DEFAULT_SESSION_ID) {
+        if (sessionManager.sessionId != DEFAULT_SESSION_ID) {
             addSessionIdToEvent(event)
-            if (!isSessionManual) {
-                updateLastActivityTime()
+            if (!sessionManager.isSessionManual) {
+                sessionManager.updateLastActivityTime()
             }
         }
         return event
@@ -87,119 +38,12 @@ internal class SessionTrackingPlugin(
 
     private fun addSessionIdToEvent(event: Event) {
         val sessionPayload = buildJsonObject {
-            put(SESSION_ID, sessionId)
-            if (isSessionStart) {
-                updateIsSessionStartIfChanged(false)
+            put(SESSION_ID, sessionManager.sessionId)
+            if (sessionManager.isSessionStart) {
+                sessionManager.updateIsSessionStartIfChanged(false)
                 put(SESSION_START, true)
             }
         }
         event.context = event.context mergeWithHigherPriorityTo sessionPayload
-    }
-
-    private fun checkAndStartSessionOnLaunch() {
-        if (shouldStartNewSessionOnLaunch()) {
-            startSession(sessionId = generateSessionId(), isSessionManual = false)
-        }
-    }
-
-    internal fun checkAndStartSessionOnForeground() {
-        if (shouldStartNewSessionOnForeground()) {
-            startSession(sessionId = generateSessionId(), isSessionManual = false)
-        }
-    }
-
-    internal fun refreshSession() {
-        if (sessionId != DEFAULT_SESSION_ID) {
-            startSession(sessionId = generateSessionId(), shouldUpdateIsSessionManual = false)
-        }
-    }
-
-    /**
-     * Starts a new session with the given session ID.
-     *
-     * @param sessionId The session ID to start the session with.
-     * @param isSessionManual Flag to indicate if the session is manual or automatic. Defaults to `false`.
-     * @param shouldUpdateIsSessionManual Flag to indicate if the `isSessionManual` should be updated. Defaults to `true`.
-     */
-    internal fun startSession(
-        sessionId: Long,
-        isSessionManual: Boolean = false,
-        shouldUpdateIsSessionManual: Boolean = true
-    ) {
-        updateIsSessionStartIfChanged(true)
-        if (shouldUpdateIsSessionManual) {
-            updateIsSessionManualIfChanged(isSessionManual)
-        }
-        updateSessionId(sessionId)
-    }
-
-    private fun attachSessionTrackingObservers() {
-        val sessionTrackingObserver = SessionTrackingObserver(this)
-        (analytics as? AndroidAnalytics)?.addLifecycleObserver(
-            sessionTrackingObserver as ProcessLifecycleObserver
-        )
-        (analytics as? AndroidAnalytics)?.addLifecycleObserver(
-            sessionTrackingObserver as ActivityLifecycleObserver
-        )
-    }
-
-    private fun updateSessionId(sessionId: Long) {
-        sessionState.dispatch(SessionState.UpdateSessionIdAction(sessionId))
-        withSessionDispatcher {
-            sessionState.value.storeSessionId(sessionId, storage)
-        }
-    }
-
-    private fun updateIsSessionManualIfChanged(isSessionManual: Boolean) {
-        if (this.isSessionManual != isSessionManual) {
-            sessionState.dispatch(SessionState.UpdateIsSessionManualAction(isSessionManual))
-            withSessionDispatcher {
-                sessionState.value.storeIsSessionManual(isSessionManual, storage)
-            }
-        }
-    }
-
-    internal fun updateLastActivityTime() {
-        val lastActivityTime = getMonotonicCurrentTime()
-        sessionState.dispatch(SessionState.UpdateLastActivityTimeAction(lastActivityTime))
-        withSessionDispatcher {
-            sessionState.value.storeLastActivityTime(lastActivityTime, storage)
-        }
-    }
-
-    private fun updateIsSessionStartIfChanged(isSessionStart: Boolean) {
-        if (this.isSessionStart != isSessionStart) {
-            sessionState.dispatch(SessionState.UpdateIsSessionStartAction(isSessionStart))
-            withSessionDispatcher {
-                sessionState.value.storeIsSessionStart(isSessionStart, storage)
-            }
-        }
-    }
-
-    internal fun endSession() {
-        sessionState.dispatch(SessionState.EndSessionAction)
-        withSessionDispatcher {
-            sessionState.value.removeSessionData(storage)
-        }
-    }
-
-    internal fun generateSessionId(): Long {
-        return TimeUnit.MILLISECONDS.toSeconds(DateTimeUtils.getSystemCurrentTime())
-    }
-
-    private fun shouldStartNewSessionOnForeground(): Boolean {
-        return sessionId != DEFAULT_SESSION_ID && !isSessionManual && hasSessionTimedOut()
-    }
-
-    private fun shouldStartNewSessionOnLaunch(): Boolean {
-        return sessionId == DEFAULT_SESSION_ID || isSessionManual || hasSessionTimedOut()
-    }
-
-    private fun hasSessionTimedOut(): Boolean {
-        return getMonotonicCurrentTime() - lastActivityTime > sessionTimeout
-    }
-
-    private fun withSessionDispatcher(block: suspend () -> Unit) {
-        analytics.analyticsScope.launch(sessionDispatcher) { block() }
     }
 }
